+++ Examples/src/components/AppDeatilsRouters/DetailsComp.tsx
@@ -1,6 +1,6 @@
 import React, { useState, FC, useContext, useMemo, useEffect } from "react";
 import classes from "./AppDeatilsRouter.scss";
-import Chart from "../NewTabs/images/chart-d.jpg";
+import Chart from "../TopBarTabs/images/chart-d.jpg";
 import { TExamplePage } from "../AppRouter/examplePages";
 
 type TProps = {
