+++ Examples/src/components/AppTopBar/AppBarTop.tsx
@@ -1,4 +1,4 @@
-import { FC, useContext, useState, useEffect, ChangeEvent } from "react";
+import { FC, useContext, useState, useEffect } from "react";
 import AppBar from "@mui/material/AppBar";
 import Button from "@mui/material/Button";
 import Toolbar from "@mui/material/Toolbar";
