+++ Examples/src/components/GalleryItems/index.scss
