+++ Examples/src/components/TopBarTabs/images/c-left.png
