+++ Examples/src/components/TopBarTabs/images/code-type.png
